my-diary-be-15/pyproject.toml
[project]
name = "main-diary-be-15"
version = "0.1.0"
description = "my diary project"
requires-python = ">=3.14"
dependencies = [
    "aerich[toml]>=0.9.2",
    "aiomysql>=0.3.2",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "ruff>=0.14.8",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]
[tool.tortoise-orm]
models = ["app.models", "aerich.models"]
generate_schemas = true

[tool.aerich]
tortoise_orm = "app.db.base.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
